refactor(web): move endpoint consumers onto shared types
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
